Applying validations

+++ API/Controllers/ItemsController.cs
@@ -20,19 +20,23 @@ namespace API.Controllers
         {
             _dbContext = dbContext;
         }
-        [HttpGet]
+
+        [HttpGet("GetItemsByStepId/{stepId}")]
         public async Task<List<ItemDto>> GetItemByStepId(int stepId)
         {
-            var step = await _dbContext.Steps.Include(s => s.Items).Where(s => s.Id == stepId).SingleAsync();
-            var stepItems = step.Items;
+            var step = await _dbContext.Steps.Include(s => s.Items).Where(s => s.Id == stepId).SingleOrDefaultAsync();
 
-            return stepItems.Select(i => new ItemDto { Id = i.Id, Description = i.Description, Title = i.Title }).ToList();
+            return step == null ? new List<ItemDto>() : step.Items.Select(i => new ItemDto { Id = i.Id, Description = i.Description, Title = i.Title }).ToList();
         }
 
         [HttpPost("AddNewItem")]
         public async Task<int> AddNewItem(CreateNewItemDto input)
         {
-            var step = await _dbContext.Steps.Include(s => s.Items).Where(s => s.Id == input.StepId).SingleAsync();
+            var step = await _dbContext.Steps.Include(s => s.Items).Where(s => s.Id == input.StepId).SingleOrDefaultAsync();
+
+            if (step == null)
+                return -1;
+
             step.Items.Add(new Item { Description = input.Description, Title = input.Title });
 
             await _dbContext.SaveChangesAsync();
@@ -44,7 +48,10 @@ namespace API.Controllers
         [HttpPost("EditItem")]
         public async Task EditItem(EditItemDto input)
         {
-            var item = await _dbContext.Items.SingleAsync(i => i.Id == input.Id);
+            var item = await _dbContext.Items.SingleOrDefaultAsync(i => i.Id == input.Id);
+
+            if (item == null)
+                return;
 
             item.Title = input.Title;
             item.Description = input.Description;
@@ -55,7 +62,10 @@ namespace API.Controllers
         [HttpPost("RemoveItem")]
         public async Task RemoveItem(int itemId)
         {
-            var item = await _dbContext.Items.SingleAsync(i => i.Id == itemId);
+            var item = await _dbContext.Items.SingleOrDefaultAsync(i => i.Id == itemId);
+
+            if (item == null)
+                return;
 
             _dbContext.Items.Remove(item);
